requirements.txt
fastapi
uvicorn[standard]
line-bot-sdk
google.generativeai
grpcio
pillow
requests
pygame
git+https://github.com/ozgur/python-firebase
